refactor(dynamic-forms): drop barrel imports and document jump outcomes
